Clean up installed files and rename the filter registration macros.
