+++ src/NodeHelper.cpp
