+++ src/NodeHelper.hpp
